move tracing logic to separate file
